+++ src/core/component/directives/attrs/index.ts
@@ -45,7 +45,7 @@ import type { ComponentInterface } from 'core/component/interface';
 import type { DirectiveParams } from 'core/component/directives/attrs/interface';
 
 //#if runtime has dummyComponents
-import('core/component/directives/attrs/test/b-component-emitter-dummy');
+import('core/component/directives/attrs/test/b-component-directives-emitter-dummy');
 //#endif
 
 export * from 'core/component/directives/attrs/const';
@@ -325,18 +325,18 @@ ComponentEngine.directive('attrs', {
 		function parseEventListener(attrName: string, attrVal: unknown) {
 			let
 				isDOMEvent = true,
-				event = attrName.slice(1).camelize(false);
+				event = attrName.slice(1).camelize(false),
+				isOnceEvent = false;
 
 			const
 				originalEvent = event,
-				eventChunks = event.split('.');
+				eventChunks = event.split('.'),
+				flags = Object.createDict<boolean>();
 
-			const
-				flags = Object.createDict<boolean>(),
-				isOnceEvent = flags.once;
-
+			// First element is event name, we need to slice only part event modifiers
 			eventChunks.slice(1).forEach((chunk) => flags[chunk] = true);
 			event = eventChunks[0];
+			isOnceEvent = Boolean(flags.once);
 
 			if (flags.right && !event.startsWith('key')) {
 				event = 'onContextmenu';
@@ -381,6 +381,7 @@ ComponentEngine.directive('attrs', {
 				}
 			}
 
+			// FIXME: `componentCtx` is always undefined https://github.com/V4Fire/Client/issues/1336
 			if (componentCtx != null && !isDOMEvent && Object.isFunction(attrVal)) {
 				if (isOnceEvent) {
 					componentCtx.$once(originalEvent, attrVal);