+++ src/core/component/directives/attrs/test/b-component-directives-emitter-dummy/b-component-directives-emitter-dummy.ts
@@ -11,11 +11,8 @@ import iBlock, { component, system } from 'components/super/i-block/i-block';
 
 export * from 'components/super/i-block/i-block';
 
-@component({
-	functional: true
-})
-
-export default class bComponentEmitterDummy extends iBlock {
+@component({functional: true})
+export default class bComponentDirectivesEmitterDummy extends iBlock {
 	@system()
 	counter: number = 0;